+++ web-frontend/modules/database/plugin.js
@@ -145,6 +145,7 @@ import {
   BaserowLeast,
   BaserowGreatest,
   BaserowRegexReplace,
+  BaserowLink,
   BaserowTrim,
   BaserowRight,
   BaserowLeft,
@@ -152,11 +153,15 @@ import {
   BaserowFilter,
   BaserowTrunc,
   BaserowRound,
+  BaserowButton,
+  BaserowGetLinkUrl,
+  BaserowGetLinkLabel,
 } from '@baserow/modules/database/formula/functions'
 import {
   BaserowFormulaArrayType,
   BaserowFormulaBooleanType,
   BaserowFormulaCharType,
+  BaserowFormulaLinkType,
   BaserowFormulaDateIntervalType,
   BaserowFormulaDateType,
   BaserowFormulaInvalidType,
@@ -421,6 +426,11 @@ export default (context) => {
   app.$registry.register('formula_function', new BaserowFilter(context))
   app.$registry.register('formula_function', new BaserowTrunc(context))
   app.$registry.register('formula_function', new BaserowRound(context))
+  // Link functions
+  app.$registry.register('formula_function', new BaserowLink(context))
+  app.$registry.register('formula_function', new BaserowButton(context))
+  app.$registry.register('formula_function', new BaserowGetLinkUrl(context))
+  app.$registry.register('formula_function', new BaserowGetLinkLabel(context))
 
   // Formula Types
   app.$registry.register('formula_type', new BaserowFormulaTextType(context))
@@ -439,6 +449,7 @@ export default (context) => {
     'formula_type',
     new BaserowFormulaSingleSelectType(context)
   )
+  app.$registry.register('formula_type', new BaserowFormulaLinkType(context))
 
   // File preview types
   app.$registry.register('preview', new ImageFilePreview(context))